requirements.txt
alembic==0.8.8
asn1crypto==0.22.0
attrs==17.2.0
autobahn==0.10.4
Automat==0.6.0
bcrypt==3.1.1
blinker==1.4
cffi==1.8.3
click==6.6
constantly==15.1.0
cryptography==2.0.3
cssmin==0.2.0
cycler==0.10.0
enum34==1.1.6
Flask==0.11.1
Flask-Assets==0.12
Flask-Bcrypt==0.7.1
Flask-Caching==1.0.1
Flask-DebugToolbar==0.10.0
Flask-Login==0.3.2
Flask-Migrate==2.0.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
Flask-OAuthlib==0.9.4
functools32==3.2.3.post2
gevent==1.2.2
greenlet==0.4.12
hyperlink==17.3.1
idna==2.6
ImageHash==3.4
incremental==17.5.0
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.8
jsmin==2.2.1
Mako==1.0.4
MarkupSafe==0.23
matplotlib==2.0.2
numpy==1.13.1
olefile==0.44
openface==0.2.1
Pillow==4.2.1
pyasn1==0.3.3
pyasn1-modules==0.1.1
pycparser==2.14
pycrypto==2.6.1
pyOpenSSL==17.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
python-editor==1.0.1
pytz==2017.2
PyWavelets==0.5.2
redis==2.10.5
scikit-learn >= 0.17, < 0.18
scipy==0.19.1
service-identity==17.0.0
simplejson==3.8.2
six==1.10.0
SQLAlchemy==1.0.12
subprocess32==3.2.7
Twisted==17.5.0
txaio==2.8.2
webassets==0.12.0
websocket-client==0.44.0
Werkzeug==0.11.4
WTForms==2.1
zope.interface==4.4.2
